Sets estimated row height to prevent jiggling list
The code before this patch would jiggle if you set the font size to a high value
and also happened to set a larger cell size. Upon investigation, this is
caused by the default row estimation behavior of UITableView. This behavior
is to return estimated values until the tableView is onscreen. By setting the
row estimate to our theme height, the contentSize returned matches the expected
size upon screen.

This makes the layout math work correctly, thereby stopping the jiggle you can
see.